+++ server.ts
@@ -1,5 +1,7 @@
-const app = require('./src/app');
-const { PORT,NODE_ENV } = require('./config');
+import app from './src/app';
+import config from './config';
+
+const { PORT, NODE_ENV } = config;
 
 const server = app.listen(PORT, () => {
 	console.log(`Application running on port: ${PORT} in ${NODE_ENV} mode`);
