+++ src/routes/note.routes.ts
@@ -110,4 +110,4 @@ router.delete('/notes/:id', async (req: Request, res: Response) => {
   }
 });
 
-module.exports = router;
+export default router;
